refactor: move LogEntryFactory from api app to common app
refs NS-165 (put related prune_django_admin_log & tests to common app)
